src/main/java/com/gojek/parkinglot/Car.java
package com.gojek.parkinglot;

import java.util.Objects;

public final class Car {
    private final String registrationNumber;
    private final String colour;

    public Car(String registrationNumber, String colour) {
        this.registrationNumber = registrationNumber;
        this.colour = colour;
    }

    public String getRegistrationNumber() {
        return registrationNumber;
    }

    public String getColour() {
        return colour;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Car car = (Car) o;
        return Objects.equals(registrationNumber, car.registrationNumber) &&
                Objects.equals(colour, car.colour);
    }
}
